refactor: Naive move of event register/audio manager to FWGameLib
