+++ Assets/FWGameLib/Common/AudioSystem/FWGLAudioSourceExtensions.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace FWGameLib.Common.AudioSystem
+{
+    public static class FWGLAudioSourceExtensions
+    {
+        public static void AssignVolume(this AudioSource source, float audioTypeVolume, float volume)
+        {
+            source.volume = audioTypeVolume * volume;
+        }
+    }
+}
