--- Assets/_Project/Scripts/Audio/AudioType.cs
@@ -1,7 +1,0 @@
-public enum AudioType
-{
-    MASTER,
-    MUSIC,
-    SFX,
-    VOICE_LINES
-}
