--- Assets/_Project/Scripts/Audio/SoundFinishedEvent.cs
@@ -1,16 +1,0 @@
-using FWGameLib.Common.EventSystem;
-
-public class SoundFinishedEvent : IEvent
-{
-    public Sounds Sound;
-    
-    public SoundFinishedEvent(Sounds sound)
-    {
-        Sound = sound;
-    }
-    
-    public void Invoke()
-    {
-        ScaleGame.Instance.EventRegister.InvokeSoundFinishedEvent(this);
-    }
-}
