--- Assets/_Project/Scripts/Audio/StopSoundEvent.cs
@@ -1,16 +1,0 @@
-using FWGameLib.Common.EventSystem;
-
-public class StopSoundEvent : IEvent
-{
-    public Sounds SoundName;
-
-    public StopSoundEvent(Sounds soundName)
-    {
-        SoundName = soundName;
-    }
-
-    public void Invoke()
-    {
-        ScaleGame.Instance.EventRegister.InvokeStopSoundEvent(this);
-    }
-}
