--- Assets/_Project/Scripts/EventSystem.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3e9d8bbbe4de4987a85494d3d4d2eb47
-timeCreated: 1723930009
