+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/0-PreTutorial/PreTutorialState.cs
@@ -1,4 +1,6 @@
+using FWGameLib.Common.AudioSystem.Event;
 using FWGameLib.Common.StateMachine;
+using FWGameLib.InProject.AudioSystem;
 
 public class PreTutorialState : IState
 {
