+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/Gameplay/3-BossPhase1/BossPhaseState.cs
@@ -1,4 +1,6 @@
+using FWGameLib.Common.AudioSystem.Event;
 using FWGameLib.Common.StateMachine;
+using FWGameLib.InProject.AudioSystem;
 using UnityEngine;
 
 public class BossPhaseState : IState
