+++ Assets/_Project/Scripts/GameLoop/GameState/InGameState/InGameState.cs
@@ -1,5 +1,7 @@
 using System;
+using FWGameLib.Common.AudioSystem.Event;
 using FWGameLib.Common.StateMachine;
+using FWGameLib.InProject.AudioSystem;
 using UnityEngine;
 using UnityEngine.SceneManagement;
