+++ Assets/_Project/Scripts/GameLoop/GameState/IntroState/IntroState.cs
@@ -1,5 +1,7 @@
 using System.Collections;
+using FWGameLib.Common.AudioSystem.Event;
 using FWGameLib.Common.StateMachine;
+using FWGameLib.InProject.AudioSystem;
 using UnityEngine;
 using UnityEngine.SceneManagement;
