+++ Assets/_Project/Scripts/UI/UISliderHelper.cs
@@ -1,3 +1,4 @@
+using FWGameLib.InProject.AudioSystem;
 using UnityEngine;
 using UnityEngine.EventSystems;
