+++ Assets/_Project/Scripts/UI/UIVolumeSlider.cs
@@ -1,9 +1,12 @@
+using FWGameLib.Common.AudioSystem;
+using FWGameLib.Common.AudioSystem.Event;
 using UnityEngine;
+using UnityEngine.Serialization;
 using UnityEngine.UI;
 
 public class UIVolumeSlider : MonoBehaviour
 {
-    public AudioType audioType;
+    [FormerlySerializedAs("audioType")] public AudioVolumeType audioVolumeType;
 
     void Start()
     {
@@ -16,11 +19,11 @@ public class UIVolumeSlider : MonoBehaviour
     private void OnEnable()
     {
         Slider slider = GetComponent<Slider>();
-        slider.value = ScaleGame.Instance.Audio.VolumeValues[audioType];
+        slider.value = ScaleGame.Instance.Audio.VolumeValues[audioVolumeType];
     }
 
     private void OnValueChange(float newValue)
     {
-        new ChangeVolumeEvent(audioType, newValue).Invoke();
+        new ChangeVolumeEvent(audioVolumeType, newValue).Invoke();
     }
 }
